chore: remove node-api optimism feature
