+++ bin/reth/Cargo.toml
@@ -150,7 +150,6 @@ optimism = [
     "reth-payload-builder/optimism",
     "reth-optimism-payload-builder/optimism",
     "reth-ethereum-payload-builder/optimism",
-    "reth-node-api/optimism",
     "dep:reth-node-optimism",
     "reth-node-core/optimism",
 ]
