+++ crates/node-api/Cargo.toml
@@ -23,6 +23,3 @@ serde.workspace = true
 [dev-dependencies]
 # for examples
 reth-payload-builder.workspace = true
-
-[features]
-optimism = []
