+++ crates/node-api/src/engine/mod.rs
@@ -215,44 +215,6 @@ pub fn validate_payload_timestamp(
     Ok(())
 }
 
-#[cfg(feature = "optimism")]
-/// Validates the presence of the `withdrawals` field according to the payload timestamp.
-///
-/// After Canyon, withdrawals field must be [Some].
-/// Before Canyon, withdrawals field must be [None];
-///
-/// Canyon activates the Shanghai EIPs, see the Canyon specs for more details:
-/// <https://github.com/ethereum-optimism/optimism/blob/ab926c5fd1e55b5c864341c44842d6d1ca679d99/specs/superchain-upgrades.md#canyon>
-pub fn optimism_validate_withdrawals_presence(
-    chain_spec: &ChainSpec,
-    version: EngineApiMessageVersion,
-    timestamp: u64,
-    has_withdrawals: bool,
-) -> Result<(), AttributesValidationError> {
-    let is_shanghai = chain_spec.fork(Hardfork::Canyon).active_at_timestamp(timestamp);
-
-    match version {
-        EngineApiMessageVersion::V1 => {
-            if has_withdrawals {
-                return Err(AttributesValidationError::WithdrawalsNotSupportedInV1)
-            }
-            if is_shanghai {
-                return Err(AttributesValidationError::NoWithdrawalsPostShanghai)
-            }
-        }
-        EngineApiMessageVersion::V2 | EngineApiMessageVersion::V3 => {
-            if is_shanghai && !has_withdrawals {
-                return Err(AttributesValidationError::NoWithdrawalsPostShanghai)
-            }
-            if !is_shanghai && has_withdrawals {
-                return Err(AttributesValidationError::HasWithdrawalsPreShanghai)
-            }
-        }
-    };
-
-    Ok(())
-}
-
 /// Validates the presence of the `withdrawals` field according to the payload timestamp.
 /// After Shanghai, withdrawals field must be [Some].
 /// Before Shanghai, withdrawals field must be [None];
@@ -340,8 +302,8 @@ pub fn validate_parent_beacon_block_root_presence(
     Ok(())
 }
 
-/// Validates the presence or exclusion of fork-specific fields based on the payload attributes
-/// and the message version.
+/// Validates the presence or exclusion of fork-specific fields based on the ethereum payload
+/// attributes and the message version.
 pub fn validate_version_specific_fields<Type>(
     chain_spec: &ChainSpec,
     version: EngineApiMessageVersion,
@@ -351,31 +313,6 @@ where
     Type: PayloadAttributes,
 {
     validate_withdrawals_presence(
-        chain_spec,
-        version,
-        payload_or_attrs.timestamp(),
-        payload_or_attrs.withdrawals().is_some(),
-    )?;
-    validate_parent_beacon_block_root_presence(
-        chain_spec,
-        version,
-        payload_or_attrs.timestamp(),
-        payload_or_attrs.parent_beacon_block_root().is_some(),
-    )
-}
-
-#[cfg(feature = "optimism")]
-/// Validates the presence or exclusion of fork-specific fields based on the payload attributes
-/// and the message version.
-pub fn optimism_validate_version_specific_fields<Type>(
-    chain_spec: &ChainSpec,
-    version: EngineApiMessageVersion,
-    payload_or_attrs: PayloadOrAttributes<'_, Type>,
-) -> Result<(), AttributesValidationError>
-where
-    Type: PayloadAttributes,
-{
-    optimism_validate_withdrawals_presence(
         chain_spec,
         version,
         payload_or_attrs.timestamp(),
