+++ crates/node-api/src/lib.rs
@@ -11,8 +11,6 @@
 ///
 /// Notably contains the [EngineTypes] trait and implementations for ethereum mainnet types.
 pub mod engine;
-#[cfg(feature = "optimism")]
-pub use engine::optimism_validate_version_specific_fields;
 pub use engine::{
     validate_payload_timestamp, validate_version_specific_fields, validate_withdrawals_presence,
     AttributesValidationError, BuiltPayload, EngineApiMessageVersion, EngineTypes,
